FIX:move v2 tests to separate module
